Fixed numeric conversion warnings

+++ include/bridge.h
@@ -27,7 +27,7 @@ class SupportLayer;
  * \param supported_regions Pre-computed regions that the support layer would
  * support.
  */
-int bridgeAngle(const Settings& settings, const Polygons& skin_outline, const SliceDataStorage& storage, const unsigned layer_nr, const unsigned bridge_layer, const SupportLayer* support_layer, Polygons& supported_regions);
+double bridgeAngle(const Settings& settings, const Polygons& skin_outline, const SliceDataStorage& storage, const unsigned layer_nr, const unsigned bridge_layer, const SupportLayer* support_layer, Polygons& supported_regions);
 
 }//namespace cura
